fix about page + a comment for AdminUserController
