+++ ProiectII/Controllers/HomeController.cs
@@ -28,6 +28,11 @@ namespace ProiectII.Controllers
             return View();
         }
 
+        public IActionResult About()
+        {
+            return View();
+        }
+
         [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
         public IActionResult Error()
         {
